+++ runner/steps.ts
@@ -51,7 +51,6 @@ export async function loadPlugins(context: Context): Promise<Plugin[]> {
 export async function configure(context: Context): Promise<void> {
   context.emit('log:debug', 'step: configure');
   const options = context.options;
-  _.defaults(options, config.defaults());
 
   await config.expand(context);
 
@@ -81,26 +80,27 @@ export async function prepare(context: Context): Promise<void> {
 
 export async function runTests(context: Context): Promise<void> {
   context.emit('log:debug', 'step: runTests');
-  const failed = false;
 
   const result = runBrowsers(context);
   const runners = result.runners;
   context._testRunners = runners;
 
-  context._socketIOServer = socketIO(context._httpServer);
-  context._socketIOServer.on('connection', function(socket) {
-    context.emit('log:debug', 'Test client opened sideband socket');
-    socket.on('client-event', function(data: ClientMessage<any>) {
-      const runner = runners[data.browserId];
-      if (!runner) {
-        throw new Error(
-            `Unable to find browser runner for ` +
-            `browser with id: ${data.browserId}`);
-      }
-      runner.onEvent(data.event, data.data);
+  context._socketIOServers = context._httpServers.map((httpServer) => {
+    const socketIOServer = socketIO(httpServer);
+    socketIOServer.on('connection', function(socket) {
+      context.emit('log:debug', 'Test client opened sideband socket');
+      socket.on('client-event', function(data: ClientMessage<any>) {
+        const runner = runners[data.browserId];
+        if (!runner) {
+          throw new Error(
+              `Unable to find browser runner for ` +
+              `browser with id: ${data.browserId}`);
+        }
+        runner.onEvent(data.event, data.data);
+      });
     });
+    return socketIOServer;
   });
-
 
   await result.completionPromise;
 }
@@ -135,22 +135,41 @@ function runBrowsers(context: Context) {
   const errors: any[] = [];
 
   const promises: Promise<void>[] = [];
-  const runners = options.activeBrowsers.map(function(browser, id) {
-    // Needed by both `BrowserRunner` and `CliReporter`.
-    browser.id = id;
-    _.defaults(browser, options.browserOptions);
 
-    const runner = new BrowserRunner(context, browser, options);
-    promises.push(runner.donePromise.then(
-        () => {
-          context.emit('log:debug', browser, 'BrowserRunner complete');
-        },
-        (error) => {
-          context.emit('log:debug', browser, 'BrowserRunner complete');
-          errors.push(error);
-        }));
-    return runner;
-  });
+  const runners: BrowserRunner[] = [];
+  let id = 0;
+  for (const originalBrowserDef of options.activeBrowsers) {
+    let waitFor: undefined|Promise<void> = undefined;
+    for (const server of options.webserver._servers) {
+      // Needed by both `BrowserRunner` and `CliReporter`.
+      const browserDef = _.clone(originalBrowserDef);
+      browserDef.id = id++;
+      browserDef.variant = server.variant;
+      _.defaults(browserDef, options.browserOptions);
+
+      const runner =
+          new BrowserRunner(context, browserDef, options, server.url, waitFor);
+      promises.push(runner.donePromise.then(
+          () => {
+            context.emit('log:debug', browserDef, 'BrowserRunner complete');
+          },
+          (error) => {
+            context.emit('log:debug', browserDef, 'BrowserRunner complete');
+            errors.push(error);
+          }));
+      runners.push(runner);
+      if (browserDef.browserName === 'safari') {
+        // Control to Safari must be serialized. We can't launch two instances
+        // simultaneously, because security lol.
+        // https://webkit.org/blog/6900/webdriver-support-in-safari-10/
+        waitFor = runner.donePromise.catch(() => {
+          // The next runner doesn't care about errors, just wants to know when
+          // it can start.
+          return undefined;
+        });
+      }
+    }
+  }
 
   return {
     runners,